+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/sourcecontext/CollectSourcesTask.kt
@@ -67,11 +67,19 @@ internal class SourceCollector {
                             .removePrefix(File.separator)
                     val targetFile = outDir.resolve(File(relativePath))
                     if (sourceFile.isFile) {
-                        SentryPlugin.logger.debug {
-                            "Copying file ${sourceFile.absolutePath} " +
-                                "to ${targetFile.absolutePath}"
+                        if (relativePath.isBlank()) {
+                            /* ktlint-disable max-line-length */
+                            SentryPlugin.logger.debug {
+                                "Skipping ${sourceFile.absolutePath} as the plugin was unable to determine a relative path for it."
+                            }
+                            /* ktlint-enable max-line-length */
+                        } else {
+                            SentryPlugin.logger.debug {
+                                "Copying file ${sourceFile.absolutePath} " +
+                                    "to ${targetFile.absolutePath}"
+                            }
+                            sourceFile.copyTo(targetFile, true)
                         }
-                        sourceFile.copyTo(targetFile, true)
                     }
                 }
             } else {