+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/util/artifacts.kt
@@ -1,13 +1,27 @@
+/*
+ * Adapted from https://github.com/autonomousapps/dependency-analysis-android-gradle-plugin/blob/10c4fc09519beedd5583db866e4ffa28a57998f1/src/main/kotlin/com/autonomousapps/internal/artifactViews.kt
+ *
+ * Copyright 2019 Anthony Robalik
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.sentry.android.gradle.util
 
 import org.gradle.api.artifacts.ArtifactView
 import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.component.ModuleComponentIdentifier
 import org.gradle.api.attributes.Attribute
-
-/**
- * Adapted from https://gist.github.com/autonomousapps/f0133e58a612b6837f3f4f6554337035
- */
 
 private val attributeKey: Attribute<String> = Attribute.of("artifactType", String::class.java)